libc/stdlib: Add abort(); change malloc()
New file(s):

./libc/stdlib/abort.c
./libc/stdlib/files.mk

Chagend file(s):

./libc/stdlib/free.c
./libc/stdlib/malloc.c
./libc/stdlib/realloc.c

+++ libc/stdlib/malloc.c
@@ -5,7 +5,7 @@
  * Project  : PINE64 ROCK64 Bare-Metal
  * Author   : Copyright (C) 2021 Johannes Krottmayer <krjdev@gmail.com>
  * Created  : 2021-01-16
- * Modified : 
+ * Modified : 2021-01-26
  * Revised  : 
  * Version  : 0.1.0.0
  * License  : ISC (see LICENSE.txt)
@@ -17,10 +17,9 @@
  */
 
 #include <stdlib.h>
+#include <heap.h>
 
-#include "heap.h"
-
-void *malloc(int len)
+void *malloc(size_t len)
 {
     if (len < 1)
         return NULL;
